quello fatto oggi assieme

+++ app/src/main/java/com/example/room/fragments/steps/RecordsAdapter.kt
@@ -80,9 +80,9 @@ class RecordsAdapter(private val recordsViewModel: RecordsViewModel) : ListAdapt
             val currentGoal = recordsViewModel.userGoal.value?.find { it.userId == 1 }
 
             if (currentGoal != null) {
-                progressBarSteps.progress = Helpers.calculatePercentage(countSteps.toInt(),currentGoal.steps )
-                progressBarCalories.progress = Helpers.calculatePercentage(countSteps.toInt(),currentGoal.calories )
-                progressBarDistance.progress = Helpers.calculatePercentage(countSteps.toInt(),currentGoal.distance.toInt())
+                progressBarSteps.progress = Helpers.calculatePercentage(countSteps.toInt(),currentGoal.steps.toDouble() )
+                progressBarCalories.progress = Helpers.calculatePercentage(countSteps.toInt(),currentGoal.calories.toDouble() )
+                progressBarDistance.progress = Helpers.calculatePercentage(countSteps.toInt(),currentGoal.distance.toDouble())
             }
